feat: add firebase db to app
