+++ src/pages/_app.tsx
@@ -1,10 +1,6 @@
-import '@/styles/globals.css'
-import type { AppProps } from 'next/app'
+import "@/styles/globals.css";
+import type { AppProps } from "next/app";
 
 export default function App({ Component, pageProps }: AppProps) {
-  return (
-    // <AuthProvider>
-      <Component {...pageProps} />
-    // </AuthProvider>
-  );
+  return <Component {...pageProps} />;
 }
